LayoutTests/imported/w3c/web-platform-tests/worklets/audio-worklet-credentials.https-expected.txt

CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin. Status code: 500

FAIL Importing a same-origin script with the default WorkletOptions should send the credentials promise_test: Unhandled rejection with value: object "AbortError: Importing a module script failed."
PASS Importing a remote-origin script with the default WorkletOptions should not send the credentials
FAIL Importing a same-origin script with credentials=omit should not send the credentials promise_test: Unhandled rejection with value: object "AbortError: Importing a module script failed."
PASS Importing a remote-origin script with credentials=omit should not send the credentials
FAIL Importing a same-origin script with credentials=same-origin should send the credentials promise_test: Unhandled rejection with value: object "AbortError: Importing a module script failed."
PASS Importing a remote-origin script with credentials=same-origin should not send the credentials
FAIL Importing a same-origin script with credentials=include should send the credentials promise_test: Unhandled rejection with value: object "AbortError: Importing a module script failed."
FAIL Importing a remote-origin script with credentials=include should send the credentials promise_test: Unhandled rejection with value: object "AbortError: Cross-origin script load denied by Cross-Origin Resource Sharing policy."
